+++ pack.py
@@ -43,15 +43,25 @@ def exec_cmd(*args):
 CUR_DIR = dirname(__file__)
 
 # The root directory
-ROOT = dirname(CUR_DIR)
+# ROOT = dirname(CUR_DIR)
 ROOT = CUR_DIR
 
-# Load main package name
+# Module name
+MODULE_NAME = ''
+
+# Version
+VERSION = ''
+
+# Load module name and version
 with open(join(ROOT, 'setup.py')) as f:
     for line in f.readlines():
         if line.startswith('NAME'):
             line = line.strip().replace(' ', '').replace('"', '').replace("'", '')
-            MAIN_PACKAGE_NAME = line.split('=')[-1]
+            MODULE_NAME = line.split('=')[-1]
+        elif line.startswith('VERSION'):
+            line = line.strip().replace(' ', '').replace('"', '').replace("'", '')
+            VERSION = line.split('=')[-1]
+        if MODULE_NAME and VERSION:
             break
 
 
@@ -70,25 +80,39 @@ def remove_paths(paths):
 
 
 def clean():
-    """Remove old version."""
-    remove_paths([join(ROOT, 'build'), join(ROOT, 'dist'), join(ROOT, '{}.egg-info'.format(MAIN_PACKAGE_NAME))])
+    """Remove the old version."""
+    remove_paths([join(ROOT, 'build'), join(ROOT, 'dist'), join(ROOT, '{}.egg-info'.format(MODULE_NAME))])
 
 
 def pack():
     """Package new version."""
     print(exec_cmd(executable, join(ROOT, 'setup.py'), 'sdist', 'bdist_wheel'))
     for p in listdir(CUR_DIR):
-        if p in ('build', 'dist', '{}.egg-info'.format(MAIN_PACKAGE_NAME)):
+        if p in ('build', 'dist', '{}.egg-info'.format(MODULE_NAME)):
             move(join(CUR_DIR, p), join(ROOT, p))
 
 
 def check():
-    """Check current version."""
+    """Check the current version."""
     print(exec_cmd('twine', 'check', join(ROOT, 'dist', '*')))
 
 
+def install():
+    """Install the current version."""
+    wheel = ''
+    for p in listdir('dist'):
+        if p.startswith('{}-{}'.format(MODULE_NAME, VERSION)) and p.endswith('.whl'):
+            wheel = join('dist', p)
+            break
+    if wheel:
+        print(exec_cmd('pip', 'uninstall', '-y', MODULE_NAME))
+        print(exec_cmd('pip', 'install', wheel))
+    else:
+        raise FileNotFoundError('wheel file does not exist')
+
+
 def upload():
-    """Upload current version to PYPI."""
+    """Upload the current version to PYPI."""
     print(exec_cmd('twine', 'upload', join(ROOT, 'dist', '*')))
 
 
@@ -96,31 +120,27 @@ def main():
     note = '''Packing tool
 
 Usage:
-
-
-\t1. pack r:\t\tRemove old version.
-
+\t1. pack d:\t\tDelete the old version.
 
 \t2. pack p:\t\tPackage new version.
 
+\t3. pack i:\t\tInstall the current version.
 
-\t3. pack c:\t\tCheck current version.
-
-
-\t4. pack u:\t\tUpload current version.
-
+\t3. pack u:\t\tUpload the current version.
 
 '''
     if 2 != len(argv):
         print(note)
         return
     operation = argv[1]
-    if 'r' == operation:
+    if 'd' == operation:
         clean()
     elif 'p' == operation:
+        clean()
         pack()
-    elif 'c' == operation:
         check()
+    elif 'i' == operation:
+        install()
     elif 'u' == operation:
         if 'y' == input('Upload to PYPI (y/n)? '):
             upload()