+++ CHANGES.md
@@ -1,4 +1,11 @@
 # Release Note
+## v1.1.2
+#### Refactor
+- Remove property "name" from "FileDownloader".
+- Complete class "UserAgentGenerator" in "MultiprocessingSpider.Utils".
+- Continue to optimize the setter method of each property. An exception will be raised if the value is invalid. "sleep_time" now can be set to 0.
+- Change the sleep strategy of subprocess, subprocess will sleep after receiving the task package to prevent multiple requests from being sent at the same time.
+___
 ## v1.1.1
 #### Bug Fixes
 - Fix "start_urls" invalidation.
@@ -14,7 +21,7 @@ ___
 - Will not resend request when "status_code" is not between 200 and 300.
 ##### a) MultiprocessingSpider
 - Rename property "handled_url_table" to "handled_urls".
-- Remove "parse" method, add "example_parse_method".
+- Remove method "parse", add "example_parse_method".
 - "User-Agent" in "web_headers" is now generated randomly.
 - Change url_table parsing order, current rule: "FIFP" (first in first parse).
 ##### b) FileDownloader